Split editor controller into focused modules
